リストのお気に入り機能

+++ src/views/List.vue
@@ -13,6 +13,9 @@
         <div v-else key="detail">
           <div class="list-menu u-mb-medium">
             <button v-if="hasSentences" class="btn btn--blue" @click="startLearning">覚える</button>
+
+            <LikeIcon v-if="loginUser" :listId="id" :userId="loginUser.id" />
+
           </div>
           <div v-if="isOwner" class="user-menu u-mb-medium">
             <router-link :to="{ name: 'listsSelect', params: id }" class="link-text">例文を選択</router-link>
@@ -42,6 +45,7 @@ import Card from '../components/Card'
 import ErrMsg from '../components/ErrMsg'
 import SentencesList from '../components/SentencesList'
 import LearningCard from '../components/LearningCard'
+import LikeIcon from '../components/LikeIcon'
 import firebase from 'firebase'
 import { fetchListById, fetchSentenceById } from '../lib/functions'
 
@@ -53,14 +57,16 @@ export default {
     return {
       list: null,
       listSentences: [],
-      isLearning: false // 学習画面の表示管理
+      isLearning: false, // 学習画面の表示管理
+      isProcessingDelete: false // リスト削除処理中の管理フラグ
     }
   },
   components: {
     Card,
     ErrMsg,
     SentencesList,
-    LearningCard
+    LearningCard,
+    LikeIcon
   },
   computed: {
     loginUser() {
@@ -79,15 +85,45 @@ export default {
   },
   methods: {
     async deleteList() {
+      // 連続クリック対策。処理中の場合は処理を抜ける。
+      if(this.isProcessingDelete) {
+        return
+      }
+
       // エラーメッセージ初期化
       this.$store.dispatch('error/clearError')
 
       try {
+        // 処理中フラグを立てる。
+        this.isProcessingDelete = true
+
+        // 当該リストに紐付く、likesコレクションのドキュメントIDを全て取得
+        const likesSnapshot = await firebase.firestore().collection('likes').where('listId', '==', this.id).get()
+        const deleteLikesIds = likesSnapshot.docs.map(doc => doc.id)
+
+        // likesドキュメントを削除
+        if(deleteLikesIds.length > 0) {
+          const deleteLikesPromises = deleteLikesIds.map(likeId => {
+            return new Promise((resolve, reject) => {
+              firebase.firestore().collection('likes').doc(likeId).delete()
+                .then(() => {
+                  resolve(true)
+                })
+                .catch((err) => {
+                  reject(err)
+                })
+            })
+          })
+
+          await Promise.all(deleteLikesPromises)
+        }
+
         // リストの削除
         await firebase.firestore().collection('lists').doc(this.id).delete()
         this.$router.push({ name: 'dashboard' })
       } catch(error) {
         this.$store.dispatch('error/setError', error)
+        this.isProcessingDelete = false
       }
     },
     startLearning() {
@@ -150,7 +186,12 @@ export default {
 
 .list-menu {
   display: flex;
+  align-items: center;
   justify-content: center;
+
+  & .btn:not(:last-child) {
+    margin-right: 4rem;
+  }
 }
 
 .user-menu {
